# docs/swagger.yaml
basePath: /
definitions:
...
    handler.ErrorDetail:
        properties:
            code:
                type: string
            message:
                type: string
        type: object
    handler.ErrorResponse:
        properties:
            error:
                $ref: '#/definitions/handler.ErrorDetail'
        type: object
...
    handler.LoginResponse:
        properties:
            access_token:
                type: string
            refresh_token:
                type: string
            user:
                $ref: '#/definitions/handler.UserSummary'
        type: object
    handler.MessageResponse:
        properties:
            message:
                type: string
        type: object
...
    handler.RegisterResponse:
        properties:
            user:
                $ref: '#/definitions/handler.UserSummary'
        type: object
...
    handler.TokenPair:
        properties:
            access_token:
                type: string
            refresh_token:
                type: string
        type: object
    handler.UserSummary:
        properties:
            email:
                type: string
            id:
                type: string
            phone:
                type: string
            role:
                type: string
            status:
                type: string
        type: object
...
    handler.forgotPasswordRequest:
        properties:
            email:
                type: string
        required:
            - email
        type: object
...
    handler.loginRequest:
        properties:
            email:
                type: string
            password:
                type: string
        required:
            - email
            - password
        type: object
    handler.logoutRequest:
        properties:
            refresh_token:
                type: string
        required:
            - refresh_token
        type: object
...
    handler.refreshRequest:
        properties:
            refresh_token:
                type: string
        required:
            - refresh_token
        type: object
    handler.registerRequest:
        properties:
            email:
                type: string
            password:
                minLength: 8
                type: string
            phone:
                type: string
        required:
            - email
            - password
            - phone
        type: object
    handler.resetPasswordRequest:
        properties:
            new_password:
                minLength: 8
                type: string
            token:
                type: string
        required:
            - new_password
            - token
        type: object
...
    handler.socialLoginRequest:
        properties:
            provider:
                type: string
            token:
                type: string
        required:
            - provider
            - token
        type: object
...
    handler.verifyCodeRequest:
        properties:
            code:
                type: string
        required:
            - code
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: 'Go + Gin ride-hailing backend: auth, rides, driver dispatch, geolocation, navigation, and places.'
    title: Ride-Hailing API
    version: 0.1.0
paths:
    /api/v1/auth/forgot-password:
        post:
            consumes:
                - application/json
            description: Triggers a password reset flow for the given email. A reset token is returned in development.
            operationId: postAuth-forgot-password
            parameters:
                - description: Forgot password request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.forgotPasswordRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Request failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Request a password reset
            tags:
                - auth
    /api/v1/auth/login:
        post:
            consumes:
                - application/json
            description: Authenticates credentials and returns access and refresh tokens.
            operationId: postAuth-login
            parameters:
                - description: Login request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.loginRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.LoginResponse'
                "401":
                    description: Invalid credentials
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Log in a user
            tags:
                - auth
    /api/v1/auth/logout:
        post:
            consumes:
                - application/json
            description: Revokes the supplied refresh token.
            operationId: postAuth-logout
            parameters:
                - description: Logout request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.logoutRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Revocation failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Log out a user
            tags:
                - auth
    /api/v1/auth/refresh:
        post:
            consumes:
                - application/json
            description: Exchanges a valid refresh token for a new token pair.
            operationId: postAuth-refresh
            parameters:
                - description: Refresh request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.refreshRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.TokenPair'
                "401":
                    description: Invalid or expired refresh token
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Refresh an access token
            tags:
                - auth
    /api/v1/auth/register:
        post:
            consumes:
                - application/json
            description: Creates a user account with email, phone, and password. The account starts in the `rider` role.
            operationId: postAuth-register
            parameters:
                - description: Registration request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.registerRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/handler.RegisterResponse'
                "409":
                    description: Account already exists
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Register a new user account
            tags:
                - auth
    /api/v1/auth/reset-password:
        post:
            consumes:
                - application/json
            description: Sets a new password using a reset token obtained from forgot-password.
            operationId: postAuth-reset-password
            parameters:
                - description: Reset password request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.resetPasswordRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "400":
                    description: Invalid or expired token
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Reset a password
            tags:
                - auth
    /api/v1/auth/social:
        post:
            consumes:
                - application/json
            description: Placeholder for OAuth-based social login. Not yet implemented.
            operationId: postAuth-social
            parameters:
                - description: Social login request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.socialLoginRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            summary: Social login (stub)
            tags:
                - auth
    /api/v1/auth/verify-email:
        post:
            consumes:
                - application/json
            operationId: postAuth-verify-email
            parameters:
                - description: Email verification request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.verifyCodeRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "400":
                    description: Invalid code
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Verify the user's email
            tags:
                - auth
    /api/v1/auth/verify-phone:
        post:
            consumes:
                - application/json
            operationId: postAuth-verify-phone
            parameters:
                - description: Phone verification request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.verifyCodeRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "400":
                    description: Invalid code
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Verify the user's phone number
            tags:
                - auth
...
securityDefinitions:
    BearerAuth:
        description: 'JWT access token. Include as `Authorization: Bearer <token>`.'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"